Update to opendaq 3.20

+++ asam_cmp_data_sink/tests/test_stream_fb.cpp
@@ -649,7 +649,7 @@ protected:
     std::shared_ptr<Packet> createEthernetPacket()
     {
         EthernetPayload ethernetPayload;
-        ethernetPayload.setData(binaryData.data(), binaryData.size());
+        ethernetPayload.setData(binaryData.data(), (uint16_t)binaryData.size());
 
         ethernetPacket = std::make_shared<Packet>();
         ethernetPacket->setPayload(ethernetPayload);
